Update code to use size_t
The`size_t` type is used to represent the size of an object.
It is defined as the return type of the sizeof operator. As an
implication size_t is used for holding an array index.

Many of the functions that would naturally return `size_t`
also return -1 when an error has occured. To hold the error
these functions still return int or long.

Being consistant with the usage of `size_t` helps reduce the
amount of casting when compiling for systems that are not
32-bit.

This is a first pass through the code to identify locations
that `int` or `long` was used when `size_t` probably should
have been used.

This is not an exastive update.

Reviewed-on: https://gerrit.iotivity.org/gerrit/26619
Tested-by: IoTivity Jenkins <jenkins-daemon@iotivity.org>

+++ port/contiki/ipadapter.c
@@ -80,7 +80,7 @@ receive(struct simple_udp_connection *c, const uip_ipaddr_t *sender_addr,
 }
 
 oc_endpoint_t *
-oc_connectivity_get_endpoints(int device)
+oc_connectivity_get_endpoints(size_t device)
 {
   (void)device;
   if (!eps) {
@@ -175,7 +175,7 @@ oc_send_buffer(oc_message_t *message)
 }
 
 int
-oc_connectivity_init(int device)
+oc_connectivity_init(size_t device)
 {
   (void)device;
   process_start(&ip_adapter_process, NULL);
@@ -183,7 +183,7 @@ oc_connectivity_init(int device)
 }
 
 void
-oc_connectivity_shutdown(int device)
+oc_connectivity_shutdown(size_t device)
 {
   (void)device;
   free_endpoints();
